Adding Authoriaztion Checks

+++ controllers/feed.js
@@ -5,6 +5,8 @@ const path = require('path')
 const { validationResult } = require('express-validator')
 
 const Post = require('../models/post')
+const User = require('../models/user')
+
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1
@@ -53,20 +55,29 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path
   const title = req.body.title
   const content = req.body.content
+  let creator
 
   const post = new Post({
     title: title,
     imageUrl: imageUrl,
     content: content,
-    creator: { name: 'sakshi' }
+    creator: req.userId
   })
 
   post.save()
     .then(result => {
-      console.log(result)
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user
+      user.posts.push(post)
+      return user.save()
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       })
     })
     .catch(err => {
@@ -132,6 +143,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized')
+        error.statusCode = 403
+        throw error
+      }
+
       post.title = title
       post.imageUrl = imageUrl
       post.content = content
@@ -165,8 +183,11 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
-      //check the logged in user 
-
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized')
+        error.statusCode = 403
+        throw error
+      }
       clearImage(post.imageUrl)
 
       return Post.findByIdAndRemove(postId)
